Update proxmox-template-create.sh
bug fixes for vmid, timezone, ipv6, balloon memory, importing disk, sshkeys

+++ proxmox-template-create.sh
@@ -10,10 +10,10 @@ usage() {
     echo "  -u, --username   Set a custom username for the cloud-init user (default: 'root')."
     echo "  -p, --pass       Set a custom password for the cloud-init user."
     echo "  -s, --sshkeys    Set SSH keys for the cloud-init user."
-    echo "  -i, --image      Specify a custom image URL for the VM template." 
+    echo "  -i, --image      Specify a custom image URL for the VM template."
     echo "  -s, --storage    Specify the storage location for the VM template (default: 'local')."
     echo "  -d, --disk-size  Specify disk size (default '32G')."
-    echo "  -t, --timezone   Set a timezone (default: 'Europe/Prague')."
+    echo "  -t, --timezone   Set a timezone (default: 'Europe/London')."
     echo "  -n, --name       Set the time of the VM"
     echo "  -c, --clean      Remove libguestfs-tools"
     echo "  -h, --help       Display this help message and exit."
@@ -24,7 +24,7 @@ usage() {
 set -e
 
 # Default values
-VMID="900"
+VMID="9000"
 USERNAME="administrator"
 PASSWORD="password"
 SSHKEYS=""
@@ -39,6 +39,7 @@ CLEAN=0
 # Parse command line arguments
 while [ "$#" -gt 0 ]; do
     case "$1" in
+           --vmid) VMID="$2"; shift ;;
         -f|--force) FORCE=1 ;;
         -u|--username) USERNAME="$2"; shift ;;
         -p|--pass) PASSWORD="$2"; shift ;;
@@ -77,7 +78,7 @@ if [ $FORCE -eq 1 ] || [ ! -f "$IMAGE_NAME" ]; then
     echo "Removing old image and downloading a new one..."
     rm -fv "$IMAGE_NAME"
     echo "Downloading Ubuntu Cloud Image..."
-    wget --inet4-only "$IMAGE_URL"
+    wget "$IMAGE_URL"
 else
     echo "Image already exists. Skipping download..."
 fi
@@ -90,7 +91,7 @@ while  qm status $VMID >/dev/null 2>&1; do
         _destroy=1
     else
         read -p "Template $VMID already exists, do you want to replace it? y/[n] " _repl
-        if [ -z "$_repl" ] || [ "$_repl" == "n" ]; then
+        if [ -z "$_repl" ] || [ "$_repl" = "n" ] || [ "$_repl" = "N" ]; then
             read -p "Enter a new template ID: " VMID
         else
             _destroy=1
@@ -127,7 +128,7 @@ virt-customize -a "$IMAGE_NAME" \
 echo "Creating VM template..."
 [ ! -z "$NAME" ] && NAME_OPT="--name"
 qm create $VMID ${NAME:+--name $NAME} --machine q35 --ostype l26 --cpu host \
-    --cores 2 --memory 1024 --balloon 8192 --onboot 1 --agent enabled=1 \
+    --cores 2 --memory 1024 --balloon 1024 --onboot 1 --agent enabled=1 \
     --net0 virtio,bridge=vmbr0,firewall=1 \
     --bios ovmf --efidisk0 "$STORAGE:0,efitype=4m" \
     --serial0 socket --vga serial0 --scsihw virtio-scsi-single \
@@ -135,7 +136,7 @@ qm create $VMID ${NAME:+--name $NAME} --machine q35 --ostype l26 --cpu host \
 # Import the image to VM and convert to QCOW2
 echo "Importing image into the VM..."
 IMPORT_OUTPUT=`qm importdisk $VMID "$IMAGE_NAME" "$STORAGE" --format qcow2 2>&1`
-DISK_NAME=`echo "$IMPORT_OUTPUT" | grep -oP "Successfully imported disk as \'\K[^']+" | sed 's/^unused0://'`
+DISK_NAME=`echo "$IMPORT_OUTPUT" | grep -i -o "imported disk '[^']*" | sed "s/imported disk '//" | sed 's/^unused0://'`
 
 # Verify disk name was captured
 if [ -z "$DISK_NAME" ]; then
@@ -157,13 +158,16 @@ qm set $VMID --boot c --bootdisk scsi0
 echo "Adding cloud-init drive..."
 qm set $VMID --scsi1 "$STORAGE:cloudinit"
 
-# Set user/password
+# Set user/password/sshkeys
 echo "Setting cloud-init user and password..."
-if [ ! -f "$SSHKEYS" ]; then
+if [ ! -f "$SSHKEYS" ] && [ -n "$SSHKEYS" ]; then
     echo "$SSHKEYS" > /tmp/sshkeys
     SSHKEYS=/tmp/sshkeys
 fi
 qm set $VMID --ciuser "$USERNAME" ${PASSWORD:+--cipassword "$PASSWORD"} ${SSHKEYS:+--sshkeys "$SSHKEYS"}
+if [ ! -f "$SSHKEYS" ] && [ -n "$SSHKEYS" ]; then
+    rm -f /tmp/sshkeys
+fi
 
 # Convert VM to template
 echo "Converting VM to template..."
